+++ node/index.d.ts
@@ -1,0 +1,2 @@
+declare const $writeFile: (filename: string, data: any | any[]) => Promise<void>;
+export { $writeFile };