+++ changelog.md
@@ -1,7 +1,10 @@
 # changelog
+## version 1.0.8
+- add node modules (writefile)
+- add params module in node and esm; not avaliable in browser
+
 ## version 1.0.7
 - remove legacy module (if using IE using the versions 1.0.3 - 1.0.6)
-- new universal module search params (node and browser)
 
 ## version 1.0.6
 - add new module intl